Add debian python sys.path workaround
Since the shebang only specifies the executable name for
env, python has to find the full path in PATH. If there is
no PATH (slurmctld runs scripts with an empty environ),
python searches for libraries from the working dir. When
slurmctld is not daemonized (-D), it does not change the
working dir away from /. So python finds /lib instead of
/usr/lib and fails to add debian's site pip package location
in /usr/local/lib/...
Our workaround is to set the working directory from systemd.