Fix config delta causing login node bounce
The addition of login_startup_scripts mysteriously made terraform start
bouncing login nodes whenever controller metadata changed. I believe
this behavior is correct; the lack of behavior before seemed to be a
bug (unconfirmed).

The only real solution is to remove the dependency between the
controller and project metadata containing configuration data. This way
the config can change but the controller will only delta on changes to
the hardware configuration, not the slurm configuration.