Fix project id in wait_for_operation for hybrid
Some util.py functions were defaulting to the global project, which is found from
google_auth. In hybrid, this might be wrong, and we want to use the project in
config.yaml. The project in google_auth is still important before we load the
config.yaml, at which point cfg.project becomes the primary.