Add slurm cluster management daemon
This substitutes the cron with the execution of slurmsync
The full cloud setup will have the daemon installed in the image
itself, and the hybrid one, terraform will be in charge of grabbing the
template from the ansible role and generating the service and timer
files, that later the user will deploy in the controller host, as
instructed in the documentation. The script setup.py will now do a
systemctl enable --now instead of creating the cron job.

Issue #219